portwatcher/pyproject.toml
[project]
name = "portwatcher"
version = "0.1.0"
description = "Django-based port scanning application for monitoring hosts"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.1",
    "python-nmap>=0.7.1",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "django-stubs>=5.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["portwatcher", "scanner"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "portwatcher.settings"
